refactor: move 'mocks.ts' to /src/mocks folder
